pyproject.toml
[tool.poetry]
name = "autoshard-scaler"
version = "0.1.0"
description = ""
authors = ["Fourat <fourat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
